+++ src/engine.rs
@@ -1,2 +1,25 @@
 mod _edit;
 mod _play;
+
+use crate::*;
+
+impl Engine<Edit> {
+    pub fn to_play(&self) -> Engine<Play> {
+        Engine {
+            _mode: PhantomData,
+            name: self.name.clone(),
+            description: self.description.clone(),
+            dig_inputs: self.dig_inputs.clone(),
+            gut_max_pressed: self.gut_max_pressed.clone(),
+            gut_min_pressed: self.gut_min_pressed.clone(),
+            gut_radio_mode: self.gut_radio_mode.clone(),
+            gut_holds: self.gut_holds.clone(),
+            guts: self.guts.clone(),
+            v_multi: self.v_multi.clone(),
+            f_multi: self.f_multi.clone(),
+            c_multi: self.c_multi.clone(),
+            index_delta_bool: self.index_delta_bool.clone(),
+            extra_delta_bool: self.extra_delta_bool.clone(),
+        }
+    }
+}
